Add files via upload
You read and rand beersong code from headfirst python.
What did you learn?
In we are solving a quiz.
We will take number of beer bottles and we have to pass it one by one.
We have to print  a stament that we have passed a bottle in a decrement order.
Example: If we have 10 bottles it will go to 10 to 1.

I asked you what did you learn? You are telling me what the program does?
I understood how to implement
for loop
if else statement.
and range.
What is range?
Range is a like a covered area.
Is range a keyword? 
Yes.
Is range a builtin function?
Yes
Is range a keyword or a builtin function.
It is both.
What is difference between keyword and a function?
All the functions are keywords but all the keywords are not functions.
How can you verify the above statement?
for is a keyword but it is not function.
range is function and keyword.
What is a keyword?
Key  is word which has specific meaning.
Where can  you get a list of keywords?
By using help.
Okay we discovered range is not a keyword.
That mean my original statement was wrong.

What did you learn by this interaction?