It sorts the vowels in the program.
What did I learn?
I learnt about 'in' and 'not in'
append
What are they?
These are operators.
What is an operator?
The Operator is a special symbol or sign which does specific work.
Ok. Then What 'in' operator does?
If I say "letter in word ", this means letter is a member of word(or a part of word data list).